\name{cms_msdrg}
\alias{cms_msdrg}
\title{Import MSDRG weights}
\description{
Imports a dataset containing MS-DRG's, descriptions, associated payment weights, and length of stay for FY2022 CMS Final Rule
}
\usage{
cms_msdrg()
}
\value{
A \code{tibble::tibble}
}
\author{Alex Zajichek}
\source{https://www.cms.gov/medicare/acute-inpatient-pps/fy-2022-ipps-final-rule-home-page}
\examples{
cms_msdrg()
}  
