\docType{data}
\name{hospitals}
\alias{hospitals}
\title{Hospitals registered with Medicare}
\description{
Collection of hospitals registered with Medicare, including geocoded addresses, payments, etc. 
}
\format{A tibble::tibble}
\usage{hospitals}
\source{https://data.cms.gov/provider-data/dataset/xubh-q36u}
\keyword{datasets}