\docType{data}
\name{msdrg}
\alias{msdrg}
\title{MS-DRG information}
\format{See "Source" for link to dataset home page}
\usage{
msdrg
}
\description{
MS-DRG descriptions, attributes, weights, length of stay
}
\source{https://www.cms.gov/medicare/acute-inpatient-pps/fy-2022-ipps-final-rule-home-page}
\keyword{datasets}