\docType{data}
\name{payments}
\alias{payments}
\title{Medicare payments by hospital and MS-DRG}
\format{See "Source" for link to dataset home page}
\usage{
payments
}
\description{
Charge and payment amounts by MS-DRG for IPPS hospitals.
}
\source{https://data.cms.gov/provider-summary-by-type-of-service/medicare-inpatient-hospitals/medicare-inpatient-hospitals-by-provider-and-service}
\keyword{datasets}