\name{pdc_topics}
\alias{pdc_topics}
\title{Extract available dataset topics from the CMS Provider Data Catalog}
\description{
Extracts the topic names in which data sets are available (see \link{https://data.cms.gov/provider-data/topics}).
}
\usage{
pdc_topics()
}
\value{
A character vector
}
\author{Alex Zajichek}
