\docType{data}
\name{hospitals}
\alias{hospitals}
\title{Hospitals registered with Medicare}
\format{See "Source" for link to dataset home page}
\usage{
hospitals
}
\description{
Information associated with hospitals registered with Medicare, including geocoded addresses, payments, etc. 
}
\source{https://data.cms.gov/provider-data/dataset/xubh-q36u}
\keyword{datasets}