\docType{data}
\name{hospitals}
\alias{hospitals}
\title{Hospitals}
\format{See "Source" for link to dataset home page}
\usage{
hospitals
}
\description{
Geocoded addresses (lat/long) and other information for hospitals registered with Medicare.
}
\source{https://data.cms.gov/provider-data/dataset/xubh-q36u}
\keyword{datasets}